% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{absorbance_read}
\alias{absorbance_read}
\title{Reading absorbance data from txt and csv files}
\usage{
absorbance_read(
  absorbance_path,
  order = TRUE,
  recursive = TRUE,
  dec = NULL,
  sep = NULL,
  verbose = FALSE,
  cores = parallel::detectCores(logical = FALSE)
)
}
\arguments{
\item{absorbance_path}{directory containing absorbance data files or path to single file. See details for format of absorbance data.}

\item{order}{logical, data is ordered according to wavelength}

\item{recursive}{read files recursive, include subfolders}

\item{dec}{optional, either you set a decimal separator or the table is tested for . and ,}

\item{sep}{optional, either you set a field separator or it is tried to be determined automatically}

\item{verbose}{logical, provide more information}

\item{cores}{number of CPU cores to be used simultanuously}
}
\description{
Exact function from 'staRdom' package, extracted to prevent orphan package warnings.
Reading absorbance data from txt and csv files.
}
