% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{check_samps}
\alias{check_samps}
\title{Checks to make sure samples and metadata match}
\usage{
check_samps(meta, eemlist, blanklist, abs)
}
\arguments{
\item{meta}{metadata table with sample information, needs to have a 'unique_ID' column}

\item{eemlist}{object of class eemlist containing EEMs samples}

\item{blanklist}{object of class eemlist containing the blanks for the EEMs samples, names should be same as eemlist with "_blank" at the end}

\item{abs}{dataframe containing absorbance data corresponding to the EEMs samples}
}
\value{
a list of metadata, eemlist, blanklist, and absorbance
}
\description{
Takes the loaded EEMs and absorbance data and tries to match it to the 'unique_ID'
in the metadata, if a sample is in the metadata but not in the loaded EEMs and
absorbance it will remove that line in the metadata. If there are samples loaded
not in the metadata it will warn and remove from the loaded samples.
}
