% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_files.R
\name{clean_files}
\alias{clean_files}
\title{Renames and organizes files from default Aqualog files}
\usage{
clean_files(prjpath, meta_file, meta_sheet, zip_files = T)
}
\arguments{
\item{prjpath}{the file path of the main file directory where data is located}

\item{meta_file}{file location of metadata table, used to determine how samples were run}

\item{meta_sheet}{optional, sheet name of the metadata only required if metadata is an .xlsx file}

\item{zip_files}{will create a zip file in the file directory of all the raw, non-renamed files from the Aqualog as a backup}
}
\value{
metadata table with an added column for the unique ID
}
\description{
Takes the default files from the Aqualog (absorbance, blank, and sample)
it renames the files to the sample identifier and creates folder for each
file type in the main file directory.
}
