% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{clean_transpose}
\alias{clean_transpose}
\title{Cleanly Transpose Data}
\usage{
clean_transpose(df)
}
\arguments{
\item{df}{A dataframe that you want to transpose.}
}
\value{
A transposed dataframe.
}
\description{
Takes a dataframe and transposes it (turn the rows to columns and columns to rows).
}
