% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_files}
\alias{create_files}
\title{Create file structure}
\usage{
create_files(prjpath)
}
\arguments{
\item{prjpath}{a string indicating the project file with the data}
}
\description{
Takes a file with EEM's and absorbance data and creates subfolders to store
raw and processed data, plots, and output tables
}
