% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data_process}
\alias{data_process}
\title{A set of processed eems data}
\format{
a list of length 2:
\describe{
  \item{1}{The processed EEMs data}
  \item{2}{The processed absorbance data}
}
}
\usage{
data_process
}
\description{
A sample set of data that's been processed and clean
}
\keyword{datasets}
