% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.document_indices}
\alias{.document_indices}
\title{Generate spectral index documenation for}
\usage{
.document_indices()
}
\value{
data.frame with information documenting column lables of spectral indices spreadsheet
}
\description{
Generate spectral index documenation for
}
