% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fewsdom_ver}
\alias{.fewsdom_ver}
\title{Returns fewsdom package version loaded}
\usage{
.fewsdom_ver()
}
\value{
text string with fewsdom package version
}
\description{
Returns fewsdom package version loaded
}
