% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_process.R
\name{eem_cut2}
\alias{eem_cut2}
\title{Trim EEMs to specified emission and/or excitation wavelengths}
\usage{
eem_cut2(eem, ex, em, exact = FALSE, fill_with_na = FALSE)
}
\arguments{
\item{eem}{An object of class eem or eemlist.}

\item{ex}{numerical of length two, with range of excitation wavelengths to keep}

\item{em}{numerical of length two, with range of emission wavelengths to keep}

\item{exact}{Logical. If TRUE, only wavelengths matching em and/or ex will be removed. If FALSE, all wavelengths in the range of em and/or ex will be removed.}

\item{fill_with_na}{Logical. If TRUE, fluorescence values at specified wavelengths will be replaced with NA. If FALSE, these values will be removed.}
}
\value{
an object of class eem or eemlist
}
\description{
Modified from eemR 'eem_cut' function. The specified wavelengths are now the
ones you keep, not the ones you remove.
}
