% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{eem_name_replace}
\alias{eem_name_replace}
\title{Replace matched patterns in sample names}
\usage{
eem_name_replace(eem_list, pattern, replacement)
}
\arguments{
\item{eem_list}{data of class eemlist}

\item{pattern}{character vector containing pattern to look for.}

\item{replacement}{character vector of replacements. Has to have the same length as pattern}
}
\description{
Exact function from 'staRdom' package, extracted to prevent orphan package warnings.
Replace matched patterns in sample names.
}
