% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattering.R
\name{eem_remove_scattering2}
\alias{eem_remove_scattering2}
\title{Removes Rayleigh and Raman Scattering}
\usage{
eem_remove_scattering2(
  eem,
  up_width = 10,
  down_width = 10,
  type = "raman",
  order = 1
)
}
\arguments{
\item{eem}{An object of class eemlist or eem.}

\item{up_width}{An integer number specifying the width in nm for the cut above the emission line.}

\item{down_width}{An integer number specifying the width in nm for the cut below the emission line.}

\item{type}{A string, either "raman" or "rayleigh".}

\item{order}{An integer number, either 1 (first order) or 2 (second order).}
}
\description{
Based on 'eem_remove_scattering' function from eemR package, added functionality
to cut different amounts from above and below the emission line. This function will only remove
one kind of scattering at a time. Use on an class eem or eemlist. Should be performed
on data that is untransformed besides a blank subtraction.
}
