% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_eems}
\alias{empty_eems}
\title{Check for EEMs that are empty}
\usage{
empty_eems(eem, verbose = T)
}
\arguments{
\item{eem}{object of class eem or eemlist}

\item{verbose}{if you want to print out which samples aren't empty}
}
\value{
a vector of EEMs with empty EEMs tables (eem$x)
}
\description{
Check for EEMs that are empty
}
