% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_files.R
\name{files_rename}
\alias{files_rename}
\title{Rename Default File Names from Aqualog}
\usage{
files_rename(meta = meta, prjpath)
}
\arguments{
\item{meta}{the metadata table for the sample run}

\item{prjpath}{a string indicating the project file containing the data to process, they should not be in subfiles}
}
\description{
Takes the raw files downloaded from the Aqualog and renames them so file names are consistent
across all run types and runs for consistent analysis and processing later.
}
