% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc.R
\name{get_doc}
\alias{get_doc}
\title{Input DOC data into metadata}
\usage{
get_doc(
  doc_file,
  doc_sheet = NULL,
  doc_column,
  name_column,
  nskip = 0,
  doc_delim = "-",
  meta_file,
  meta_sheet = NULL,
  meta_delim,
  rewrite = T
)
}
\arguments{
\item{doc_file}{a string of the file path of the DOC file, can be .xlsx or .csv}

\item{doc_sheet}{a string of the sheet name of the DOC results, only required if the DOC file is an .xlsx file}

\item{doc_column}{a numeric indicating which column the DOC results are stored in in the DOC file}

\item{name_column}{a numeric indicating which column the sample/site names are stored in the DOC file}

\item{nskip}{a numeric indicating a number of lines to skip when reading in the DOC file, optional}

\item{doc_delim}{a string indicating the separation between site name and other identifiers in the DOC data, it will only keep the first piece}

\item{meta_file}{a string indicating the file path of the metadata, can be .xlsx or .csv}

\item{meta_sheet}{a string of the metadata sheet name, only required if the metadata file is an .xlsx file}

\item{meta_delim}{a string indicating the separate between site name and other identifiers in the metadata data identifier}

\item{rewrite}{A logical, if TRUE original metadata will be saved over with metadata with DOC results, if FALSE it will add "_doc_added" to the end of the table}
}
\description{
Takes data from a separate table of DOC results and puts
it in the metadata table under the correct sample. The code assumes
DOC is reported in mg/L. See 'sample_metadata' in 'fewsdom' package
for formatting.
}
\details{
The table with DOC data can be either a .csv or excel file.
}
