% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{get_indices}
\alias{get_indices}
\title{Export Absorbance and Fluorencence Indices and Peaks to an Excel File}
\usage{
get_indices(
  eem_list,
  abs_data,
  meta,
  path,
  doc_norm = "both",
  sampsascol = F,
  waves = NULL
)
}
\arguments{
\item{eem_list}{An object of class eem or eemlist.}

\item{abs_data}{A dataframe containing the absorbance data.}

\item{meta}{The metadata associated with the samples.}

\item{path}{Location to save the exported absorbance and fluorescence metrics.}

\item{doc_norm}{Either TRUE, FALSE or "both" indicating if peaks should be normalized by DOC concentrations.}

\item{sampsascol}{A logical indicating how results should be oriented, TRUE puts samples as columns, FALSE puts samples as rows.}

\item{waves}{Optional, a vector of wavelengths in nm to extract from the absorbance data.}
}
\description{
Export Absorbance and Fluorencence Indices and Peaks to an Excel File
}
