% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggeem2}
\alias{ggeem2}
\title{Plot EEM's plots as contour plots}
\usage{
ggeem2(
  eem,
  manualscale = F,
  manualmax = 1.5,
  manualmin = 0,
  nbins = 8,
  label_peaks = F,
  prec = 2,
  palette = "parula"
)
}
\arguments{
\item{eem}{An object of class eem, the data you want to plot.}

\item{manualscale}{A TRUE or FALSE indicating if you want to specify the minimum and maximum of the intensity scale manually, good for comparisons across samples.}

\item{manualmax}{Maximum value for intensity scale.}

\item{manualmin}{Minimum value used for intensity scale.}

\item{nbins}{The number of bins (and colors) used in the contour plot. Maximum of 24.}

\item{label_peaks}{A TRUE or FALSE indicating if you want the main coble peaks annotated on the plot.}

\item{prec}{An integer for the number of significant figures used for binning the intensity.}

\item{palette}{A character with the color palette to use. Currently 'parula', 'ocean.haline', 'cubicl', and 'kovesi.rainbow' from the 'pals' package are supported.}
}
\description{
Creates a pretty EEM's contour plot with the ability to manually set the intensity
scale or use an automated one based on the minimum and maximum of the EEM. Option to
add Coble peak labels to the plot (based off peaks from Coble et al. 2014.
}
