% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{list_join}
\alias{list_join}
\title{Full join of a list of data frames.}
\usage{
list_join(df_list, by)
}
\arguments{
\item{df_list}{list of data frames to by joined}

\item{by}{character vector containing information how to join data frames. Format to be according to by in full_join. Each data frame has to contain the column(s) used for joining.}
}
\description{
Exact function from 'staRdom' package, extracted to prevent orphan package warnings.
Full join of a list of data frames.
}
