% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_eems}
\alias{load_eems}
\title{Load renamed and organized Aqualog data into R}
\usage{
load_eems(prjpath)
}
\arguments{
\item{prjpath}{The file path of the main file directory where data is located}
}
\description{
Takes the Aqualog files that have been renamed and put in separate folders within
the main directory via the 'clean_files' function and loads them into R as
eemlists. It will also check to see if maximum absorbance is below 0.3.
}
\examples{
\dontrun{
  raw_eem <- load_eems(prjpath = prjpath)
}
 X <- raw_eem[[1]] #the raw eem samples
 X_blk <- raw_eem[[2]] #the matching blanks
 Sabs <- raw_eem[[3]] #the matching absorbance data
}
