% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_eems}
\alias{plot_eems}
\title{Exports EEMs contour plots to a file}
\usage{
plot_eems(
  prjpath,
  meta,
  eem,
  sing_plot = T,
  sum_plot = T,
  doc_norm = T,
  ncol = 4,
  title = "description",
  save_names = "",
  ...
)
}
\arguments{
\item{prjpath}{a string indicating the project file}

\item{meta}{the metadata associated with the samples}

\item{eem}{object of class eemlist with EEMs to be plotted}

\item{sing_plot}{logical, if TRUE will create and save a separate plot for each EEMs}

\item{sum_plot}{logical, if TRUE will create and save a plot with all the EEMs in a single plot}

\item{doc_norm}{either TRUE, FALSE. TRUE return EEMs normalized by DOC concentration}

\item{ncol}{number of columns used in summary plot with all the EEMs samples}

\item{title}{either a dataframe with samples and titles, 'description', 'sampleID', or FALSE (see details)}

\item{save_names}{optional, a character that will be added to the unique ID for plot file names}

\item{...}{arguments to pass to 'ggeems2' function}
}
\description{
Will plot EEMs using 'ggeem2' function then will export plots to specified file.
}
\details{
Use 'create_files' function before running this function to ensure file structure is correct for pre-processing.

For title argument, you can specify the title manually, by using a dataframe where the
first column in the unique ID's for the samples and the second column is the title. If 'description'
is chosen it will use the description column of the metadata. If 'sampleID' is chosen it will
use the unique identifier column from the metadata. Lastly if FALSE is chosen no titles will be added.
}
