% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_eems.R
\name{process_eems}
\alias{process_eems}
\title{Takes raw EEMs and absorbance data from Aqualog, returns cleaned, processed data
Combines all the 'fewsdom' functions to make one function to process samples.}
\usage{
process_eems(
  prjpath,
  run_date,
  doc_file,
  doc_sheet,
  doc_column = 7,
  name_column = 4,
  nskip = 3,
  doc_delim = "-",
  meta_sheet = "log",
  site_loc = c(1, 7),
  zip_files = T,
  process_file = T,
  ...
)
}
\arguments{
\item{prjpath}{path to folder with raw samples}

\item{run_date}{date the samples were run, this should correspond to the file where all the samples are stored}

\item{doc_file}{a string of the file path of the DOC file, can be .xlsx or .csv}

\item{doc_sheet}{a string of the sheet name of the DOC results, only required if the DOC file is an .xlsx file}

\item{doc_column}{a numeric indicating which column the DOC results are stored in in the DOC file}

\item{name_column}{a numeric indicating which column the sample/site names are stored in the DOC file}

\item{nskip}{a numeric indicating a number of lines to skip when reading in the DOC file, optional}

\item{doc_delim}{a string indicating the separation between site name and other identifiers in the DOC data, it will only keep the first piece}

\item{meta_sheet}{a string of the metadata sheet name, only required if the metadata file is an .xlsx file}

\item{site_loc}{a vector indicating the start and end of the site name in the metadata data identifier}

\item{zip_files}{will create a zip file in the file directory of all the raw, non-renamed files from the Aqualog as a backup}

\item{process_file}{logical, if TRUE it will put a text file in the processed data folder named 'processing_tracking'}

\item{...}{additional arguments passed to the 'get_doc', 'clean_files', 'abs_preprocess', 'load_eems', 'eem_process', 'plot_eems', 'get_indices', 'save_eems' functions}
}
\value{
saves processed EEMs as .csv and .rds files, absorbance as .csv and .rds, and metadata as .csv
creates plots for each sample and calculates indices.
}
\description{
Takes raw EEMs and absorbance data from Aqualog, returns cleaned, processed data
Combines all the 'fewsdom' functions to make one function to process samples.
}
