% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw_eem}
\alias{raw_eem}
\title{A set of raw eems data}
\format{
a list of length 3:
\describe{
  \item{1}{The raw EEMs samples}
  \item{2}{The blanks associated with the EEMs}
  \item{3}{The absorbance data associated with the EEMs samples}
}
}
\usage{
raw_eem
}
\description{
A sample set of data that's been uploaded for testing functions
}
\keyword{datasets}
