% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_eems.R
\name{save_eems}
\alias{save_eems}
\title{Save EEMs and absorbance data to file}
\usage{
save_eems(eem_list, abs_data, meta, prjpath)
}
\arguments{
\item{eem_list}{object of class eemlist or eem}

\item{abs_data}{a dataframe containing the absorbance data}

\item{meta}{the metadata associated with the samples}

\item{prjpath}{location to save the exported EEMs and absorbance files}
}
\description{
Saves absorbance as a .csv file and EEMs as individual .csv files and .rds file
for use in future analysis.
}
\details{
Use 'clean_files' function before running this function to ensure file structure is correct for pre-processing.
}
