% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_files.R
\name{clean_files}
\alias{clean_files}
\title{Renames and organizes files from default Aqualog files}
\usage{
clean_files(prjpath, meta_file, meta_sheet = NULL, ...)
}
\arguments{
\item{prjpath}{the file path of the main file directory where data is located}

\item{meta_file}{file location of metadata table, used to determine how samples were run}

\item{meta_sheet}{optional, sheet name of the metadata only required if metadata is an .xlsx file}

\item{...}{arguments to pass down to functions with 'clean_files'}
}
\value{
metadata table with an added column for the unique ID
}
\description{
Takes the default files from the Aqualog (absorbance, blank, and sample)
it renames the files to the sample identifier and creates folder for each
file type in the main file directory.
}
\examples{
\dontrun{
meta <- clean_files(prjpath=prjpath, meta_file=meta_file,
meta_sheet = meta_sheet)
}

}
