% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattering.R
\name{eem_remove_scattering2}
\alias{eem_remove_scattering2}
\title{Removes rayleigh and raman scattering}
\usage{
eem_remove_scattering2(
  eem,
  type = "raman",
  order = 1,
  up_width = 10,
  down_width = 10
)
}
\arguments{
\item{eem}{an object of class eemlist or eem}

\item{type}{a string, either "raman" or "rayleigh"}

\item{order}{an integer number, either 1 (first order) or 2 (second order)}

\item{up_width}{an integer number specifying the width in nm for the cut above the emission line}

\item{down_width}{an integer number specifying the width in nm for the cut below the emission line}
}
\description{
Based on 'eem_remove_scattering' function from eemR package, added functionality
to cut different amounts from above and below the emission line. This function will only remove
one kind of scattering at a time. Use on an class eem or eemlist. Should be performed
on data that is untransformed besides a blank subtraction.
}
