% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_eems}
\alias{load_eems}
\title{Load renamed and organized Aqualog data into R}
\usage{
load_eems(prjpath)
}
\arguments{
\item{prjpath}{The file path of the main file directory where data is located}
}
\description{
Takes the Aqualog files that have been renamed and put in separate folders within
the main directory via the 'clean_files' function and loads them into R as
eemlists. It will also check to see if maximum absorbance is below 0.3.
}
