% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_eems}
\alias{plot_eems}
\title{Plot EEM's plots as contour plots}
\usage{
plot_eems(
  prjpath,
  meta,
  eem,
  sing_plot = T,
  sum_plot = T,
  doc_norm = T,
  title = "description",
  ...
)
}
\arguments{
\item{prjpath}{a string indicating the project file}

\item{meta}{the metadata associated with the samples}

\item{eem}{object of class eemlist with EEMs to be plotted}

\item{sing_plot}{logical, if TRUE will create and save a separate plot for each EEMs}

\item{sum_plot}{logical, if TRUE will create and save a plot with all the EEMs in a single plot}

\item{doc_norm}{either TRUE, FALSE, or "both". TRUE or "both" will return EEMs normalized by DOC concentration}

\item{title}{either a vector the same length as the number of EEMs with titles, or "description" or "sampleID"}

\item{...}{arguments to pass to 'ggeems2' function}
}
\description{
Creates a pretty EEM's contour plot with the ability to manually set the intensity
scale or use an automated one based on the minimum and maximum of the EEM. Option to
add Coble peak labels to the plot (based off peaks from Coble et al. 2014.
}
